+++ webpack.config.js
@@ -37,7 +37,11 @@ module.exports = {
         loaders: [
             {
                 test: /\.less$/,
-                loader: ExtractTextPlugin.extract('style-loader', 'css-loader!less-loader')
+                use: ExtractTextPlugin.extract({
+                    fallback: "style-loader",
+                    use: "css-loader!less-loader"
+                    // publicPath: "/dist"
+                })
             },
             {
                 test: /\.tpl\.html$/,
@@ -49,7 +53,7 @@ module.exports = {
     devtool: 'inline-source-map',
     plugins: [
         new ExtractTextPlugin("cyclone.css"),
-        new webpack.optimize.CommonsChunkPlugin('vendors', 'vendors.js'),
+        new webpack.optimize.CommonsChunkPlugin({ name: 'vendors', filename: 'vendors.js' }),
         new BrowserSyncPlugin({
             host: 'localhost',
             port: 3000,