+++ webpack.config.staging.js
@@ -31,7 +31,11 @@ module.exports = {
         loaders: [
             {
                 test: /\.less$/,
-                loader: ExtractTextPlugin.extract('style-loader', 'css-loader!less-loader')
+                use: ExtractTextPlugin.extract({
+                    fallback: "style-loader",
+                    use: "css-loader!less-loader"
+                    // publicPath: "/dist"
+                })
             },
             {
                 test: /\.tpl\.html$/,
@@ -69,7 +73,7 @@ module.exports = {
             mangle: false, // was true - but fails because of injection issues since new components
             // TODO: fix this, or update webpack builder to maybe include it directly automatically
         }),
-        new webpack.optimize.CommonsChunkPlugin('vendors', 'vendors.js'),
+        new webpack.optimize.CommonsChunkPlugin({ name: 'vendors', filename: 'vendors.js' }),
         new webpack.optimize.DedupePlugin(),
         new webpack.optimize.AggressiveMergingPlugin(),
         new FileChanger({